Changed Transaction.amount to DECIMAL data type

+++ server/app/db/schemas/transactions.py
@@ -1,16 +1,17 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 class Transaction(BaseModel):
     id: int
     entry_id: int
     member_id: int
-    amount: float
+    amount: Decimal
     class Config:
         orm_mode = True
 
 class TransactionCreatePayloadWithMember(BaseModel):
     member_id: int
-    amount: float
+    amount: Decimal
 
 class TransactionUpdatePayload(BaseModel):
-    amount: float
+    amount: Decimal
